Start random expression generator in Python
I implemented the beginning of the random valid and invalid mathematical function generator in Python. This spits out a csv value of the expressions and the expected result. The C program then reads that csv file, evaluates the expressions, and returns the percentage of expressions that passed the test. The C tester also writes a new file output with all of the expressions that fail the test case.